fix: bugs

+++ app/Http/Controllers/VideoController.php
@@ -2,13 +2,18 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Comment;
 use App\Models\Video;
+use App\Models\Comment;
+use App\Models\Interaction;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class VideoController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['auth:api'], ['only' => ['addComment', 'deleteComment', 'interaction']]);
+    }
     public function getVideo($slug)
     {
         $res = Video::with([
@@ -19,38 +24,50 @@ class VideoController extends Controller
             'category',
         ])->where('slug', $slug)->first();
 
+        if($res->id) Video::where('id', $res->id)->increment('views');
+
         return $res;
     }
     public function addComment(Request $request)
     {
-
         $request->validate([
-            'video_id' => 'required|integer',
+            'video_id' => 'required|integer|exists:videos,id',
             'comment' => 'required|string',
+            'reply' => 'mullable|exists:comments,id',
         ]);
 
-        Comment::create([
+        $res = Comment::create([
             'video_id' => $request->video_id,
             'user_id' => Auth::user()->id,
             'comment' => $request->comment,
+            'parent_id' => $request->reply ?? null,
         ]);
 
-        return response(['status' => 'success']);
+        return response(['status' => $res ? 'success' : 'error']);
     }
     public function deleteComment(Request $request)
     {
         $request->validate([
-            'comment_id' => 'required|integer',
+            'comment_id' => 'required|integer|exists:comments,id',
         ]);
-        Comment::where('id', $request->comment_id)->delete();
-        return response(['status' => 'success']);
+        $res = Comment::where('id', $request->comment_id)->where('user_id', Auth::user()->id)->delete();
+        return response(['status' => $res ? 'success' : 'error']);
     }
-    public function setView(Request $request)
+    public function interaction(Request $request)
     {
         $request->validate([
-            'video_id' =>'required|integer',
+            'interaction'=> 'required|in:like,dislike',
+            'video_id' =>'required|integer|exists:videos,id',
         ]);
-        Video::where('id', $request->video_id)->increment('views');
-        return response(['status' =>'success']);
+     
+        $res = false;
+        if(!Interaction::where('video_id', $request->video_id)->where('user_id', Auth::user()->id)->first()){
+            $res = Interaction::create([
+                'video_id' => $request->video_id,
+                'user_id' => Auth::user()->id,
+                'is_liked' => $request->interaction ? true : false,
+            ]);
+        }
+        return response(['status' => $res ? 'success' : 'error']);
     }
 }
